backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-openai
langchain-community
psycopg2-binary
pgvector
neo4j
sentence-transformers
transformers
torch
pypdf
python-docx
numpy
scikit-learn
aiofiles
pydantic
pydantic-settings
python-multipart
rank-bm25
google-generativeai
asyncpg
aiohttp
networkx